feat: add ShareDropdown component
